refactor: collections folder and namespace
Folloing .net example,  the collections have been moved to a folder
and a namespace.

BREAKING CHANGE: collections moved to Mirage.Collections
