+++ Assets/Mirage/Editor/InspectorHelper.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
+using Mirage.Collections;
 using UnityEngine;
 
 namespace Mirage
