+++ Assets/Mirage/Editor/NetworkBehaviourInspector.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
+using Mirage.Collections;
 using UnityEditor;
 using UnityEngine;
