+++ Assets/Mirage/Runtime/Collections.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 35d6ef79457a14d13b873eaf1cd3a83c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
