+++ Assets/Mirage/Runtime/Collections/SyncSet.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 
-namespace Mirage
+namespace Mirage.Collections
 {
     public class SyncSet<T> : ISet<T>, ISyncObject
     {
