+++ Assets/Mirage/Runtime/NetworkBehaviour.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using Cysharp.Threading.Tasks;
+using Mirage.Collections;
 using Mirage.RemoteCalls;
 using UnityEngine;
