+++ Assets/Tests/Runtime/Collections.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 316b093e8c2244fceac3b68dd197fc27
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
