+++ Assets/Tests/Runtime/Collections/SyncListClassTest.cs
@@ -1,4 +1,5 @@
 using System.Linq;
+using Mirage.Collections;
 using NUnit.Framework;
 
 namespace Mirage.Tests
