+++ Assets/Tests/Runtime/Host/NetworkBehaviourTests.cs
@@ -1,3 +1,4 @@
+using Mirage.Collections;
 using NUnit.Framework;
 using UnityEngine;
