+++ Assets/Tests/Runtime/Serializers/NetworkBehaviourSerializeTest.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using Mirage.Collections;
 using NUnit.Framework;
 using UnityEngine;
